Moved leaderboard over to another file
